requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
discord==2.3.2
discord-ext-menus==1.1
discord.py==2.3.2
DiscordUtils==1.3.4
frozenlist==1.4.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
lxml==4.9.2
multidict==6.0.4
Pillow==9.4.0
pycparser==2.21
PyNaCl==1.5.0
pyreadline3==3.4.1
regex==2023.3.23
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.4.1
urllib3==2.0.4
yarl==1.9.2
